Make tests a platform, unify build
