+++ src/table.h
@@ -23,6 +23,10 @@ struct table {
 float interpolate_table_oneaxis(struct table *, float a);
 float interpolate_table_twoaxis(struct table *, float a1, float a2);
 
+#ifdef UNITTEST
+#include <check.h>
+TCase *setup_table_tests();
+#endif
 
 #endif
