+++ src/util.h
@@ -11,5 +11,10 @@ timeval_t time_from_rpm_diff(unsigned int rpm, degrees_t degrees);
 int time_in_range(timeval_t val, timeval_t t1, timeval_t t2);
 degrees_t clamp_angle(int, degrees_t);
 
+#ifdef UNITTEST
+#include <check.h>
+TCase *setup_util_tests();
+#endif
+
 #endif
